create a submodule for template_localization
