+++ spikeinterface/toolkit/postprocessing/__init__.py
@@ -2,8 +2,7 @@ from .template_tools import (get_template_amplitudes,
                              get_template_extremum_channel,
                              get_template_extremum_channel_peak_shift,
                              get_template_extremum_amplitude,
-                             get_template_channel_sparsity,
-                             compute_unit_centers_of_mass)
+                             get_template_channel_sparsity)
 
 from .template_metrics import (calculate_template_metrics, get_template_metric_names)
 
@@ -15,3 +14,6 @@ from .principal_component import (WaveformPrincipalComponent,
 from .spike_amplitudes import get_spike_amplitudes
 
 from .correlograms import compute_correlograms
+
+
+from .template_localization import compute_unit_centers_of_mass
